[#51] Greatly simplify requirements
The initial list requirements was generated with `pipreqs`. The output was
cleaned up. The versions not matching the old `requirements.txt` was then
changed to match. This was then compiled with `pip-compile requirements.in` from
the `pip-tool` package to lock the versions again. After installation the output
of `pip freeze` was compared to the old requirements.txt`. It all resulted in
quite a few bumps of versions. Mostly minor versions. It also resulted in the
removal of the following packages:

Very unrelated to this project. Not imported anywhere. Very unlikely to be used.
Should be very safe to remove.

- boto3, botocore: AWS SDK.
- s3transfer: AWS S3.
- nodeenv: For Node.js.

Compilers and lowlevel libs Likely dependency of dependency. Should be very safe
to remove.

- Cython
- distlib
- llvmlite
- sip

Tooling for packaging and development.  Should be very safe to remove.

- pip
- pre-commit
- setuptools
- virtualenv
- wheel

Tooling for test and documentation. Should not live in this file, but in a
dedicated file.

- docutils
- mock
- nose
- PyQt5-stubs: For typechecking with mypy

Some various smaller libs. They are not imported anywhere, but there could be a
medium risk they are used somewhere.

- appdirs
- cfgv
- filelock
- identify
- jmespath
- toml
- pyqtgraph